+++ test/html-to-govspeak.test.js
@@ -5,6 +5,7 @@ import htmlToGovspeak from '../src/html-to-govspeak'
 it('converts HTML to govspeak', () => {
   expect(htmlToGovspeak('<p>Hello</p>')).toEqual('Hello')
 })
+
 it("doesn't escape markdown", () => {
   const html = '<p>[Markdown](link)</p>'
   expect(htmlToGovspeak(html)).toEqual('[Markdown](link)')
@@ -61,6 +62,65 @@ it('removes image elements', () => {
   expect(htmlToGovspeak('<img src="image.jpg" alt="" />')).toEqual('')
 })
 
+it('removes title elements', () => {
+  expect(htmlToGovspeak(`<title>Title</title>`)).toEqual('')
+})
+
+it('removes script elements', () => {
+  expect(htmlToGovspeak(`<script>alert('hi')</script>`)).toEqual('')
+})
+
+it('removes noscript elements', () => {
+  expect(htmlToGovspeak(`<noscript>Enable JS</noscript>`)).toEqual('')
+})
+
+it('removes style elements', () => {
+  expect(htmlToGovspeak(`<style>p {color:red;}</style>`)).toEqual('')
+})
+
+it('removes video elements', () => {
+  const html = `
+    <video width="320" height="240" controls>
+      <source src="movie.mp4" type="video/mp4">
+      <source src="movie.ogg" type="video/ogg">
+      Fallback text
+    </video>
+  `
+
+  expect(htmlToGovspeak(html)).toEqual('')
+})
+
+it('removes audio elements', () => {
+  const html = `
+    <audio controls src="/media/examples/t-rex-roar.mp3">
+      Fallback text
+    </audio>
+  `
+
+  expect(htmlToGovspeak(html)).toEqual('')
+})
+
+it('removes object elements', () => {
+  const html = `
+    <object type="application/pdf"
+      data="/media/examples/In-CC0.pdf"
+      width="250"
+      height="200">
+      Fallback text
+    </object>
+  `
+
+  expect(htmlToGovspeak(html)).toEqual('')
+})
+
+it('removes iframe elements', () => {
+  const html = `
+    <iframe src="./file">Fallback text</iframe>
+  `
+
+  expect(htmlToGovspeak(html)).toEqual('')
+})
+
 it('converts abbr elements to references', () => {
   const html = `
     <p>
@@ -102,10 +162,6 @@ it('removes rogue br elements', () => {
   expect(htmlToGovspeak(html)).toEqual(
     'Not empty\n\nNot empty either'
   )
-})
-
-it('removes style elements', () => {
-  expect(htmlToGovspeak(`<style>p {color:red;}</style>`)).toEqual('')
 })
 
 it('extracts headers from lists', () => {
@@ -203,4 +259,16 @@ it('Maintains behaviour where a <span> </span> produces a double space', () => {
   const html = `Some text<span> </span>and some more text`
 
   expect(htmlToGovspeak(html)).toEqual('Some text  and some more text')
+})
+
+// The presence of elements preceeding text that is rendered can cause
+// preceeding whitespace. Typically this is caused by elements in the <head>
+// that are to be stripped out.
+it('strips whitespace caused by surrounding, non-visual elements', () => {
+  const html = `
+    <meta charset="utf-8">
+    <title>Title</title>
+    <p>Some text</p>
+  `
+  expect(htmlToGovspeak(html)).toEqual('Some text')
 })