+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Convert h1 headers to h2, convert h4 and h5 headers to h3 (PR #41)
 - Fix dist file containing ES2015 and regex breaking in IE11 (PR #39)
+- Remove more undesirable elements and trim output whitespace (PR #38)
 
 ## 0.1.1